move trimming out of Core and into the main project with Pages
