+++ TASVideos/Pages/Wiki/Models/WikiEditModel.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using TASVideos.Core.Services;
+using TASVideos.Models;
 
 namespace TASVideos.Pages.Wiki.Models;
